requirements.txt
annotated-types==0.7.0
anyio==4.6.0
boto3==1.35.33
botocore==1.35.33
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
motor==3.6.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
